# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mandate API",
    "description": "Autonomous CFO agent for crypto-native organizations and DAOs.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/orgs/{org_id}/contractors": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Contractors",
        "operationId": "contractors_api_v1_orgs__org_id__contractors_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "title": "Response Contractors Api V1 Orgs  Org Id  Contractors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{org_id}/payroll/upload": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Upload Payroll",
        "operationId": "upload_payroll_api_v1_orgs__org_id__payroll_upload_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Org Id"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Payroll batch",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_payroll_api_v1_orgs__org_id__payroll_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Upload Payroll Api V1 Orgs  Org Id  Payroll Upload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{org_id}/payroll/sample": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Create Sample Payroll",
        "description": "One-click demo: build a batch from the bundled sample contractor CSV.",
        "operationId": "create_sample_payroll_api_v1_orgs__org_id__payroll_sample_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Org Id"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Sample payroll (12 contractors)",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Sample Payroll Api V1 Orgs  Org Id  Payroll Sample Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{org_id}/payroll/json": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Create Payroll Json",
        "operationId": "create_payroll_json_api_v1_orgs__org_id__payroll_json_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayrollJsonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Payroll Json Api V1 Orgs  Org Id  Payroll Json Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{org_id}/payroll/batches": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "List Batches",
        "operationId": "list_batches_api_v1_orgs__org_id__payroll_batches_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "title": "Response List Batches Api V1 Orgs  Org Id  Payroll Batches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{org_id}/payroll/batches/{batch_id}": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Get Batch",
        "operationId": "get_batch_api_v1_orgs__org_id__payroll_batches__batch_id__get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Org Id"
            }
          },
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Batch Api V1 Orgs  Org Id  Payroll Batches  Batch Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{org_id}/payroll/batches/{batch_id}/propose": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Propose Batch",
        "operationId": "propose_batch_api_v1_orgs__org_id__payroll_batches__batch_id__propose_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Org Id"
            }
          },
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Propose Batch Api V1 Orgs  Org Id  Payroll Batches  Batch Id  Propose Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{org_id}/payroll/batches/{batch_id}/execute": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Execute Batch",
        "operationId": "execute_batch_api_v1_orgs__org_id__payroll_batches__batch_id__execute_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Org Id"
            }
          },
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Execute Batch Api V1 Orgs  Org Id  Payroll Batches  Batch Id  Execute Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{org_id}/transfers": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Create Transfer",
        "operationId": "create_transfer_api_v1_orgs__org_id__transfers_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Transfer Api V1 Orgs  Org Id  Transfers Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{org_id}/transactions": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "List Transactions",
        "operationId": "list_transactions_api_v1_orgs__org_id__transactions_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Org Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "title": "Response List Transactions Api V1 Orgs  Org Id  Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{org_id}/swap/quote": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Swap Quote",
        "operationId": "swap_quote_api_v1_orgs__org_id__swap_quote_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwapRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Swap Quote Api V1 Orgs  Org Id  Swap Quote Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{org_id}/yield/venues": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Yield Venues",
        "operationId": "yield_venues_api_v1_orgs__org_id__yield_venues_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Org Id"
            }
          },
          {
            "name": "asset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "USDC",
              "title": "Asset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "title": "Response Yield Venues Api V1 Orgs  Org Id  Yield Venues Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orgs/{org_id}/yield/deploy": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Yield Deploy",
        "operationId": "yield_deploy_api_v1_orgs__org_id__yield_deploy_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployYieldRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Yield Deploy Api V1 Orgs  Org Id  Yield Deploy Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_payroll_api_v1_orgs__org_id__payroll_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_payroll_api_v1_orgs__org_id__payroll_upload_post"
      },
      "DeployYieldRequest": {
        "properties": {
          "amount_usd": {
            "type": "number",
            "title": "Amount Usd"
          },
          "asset": {
            "type": "string",
            "title": "Asset",
            "default": "USDC"
          },
          "venue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          }
        },
        "type": "object",
        "required": [
          "amount_usd"
        ],
        "title": "DeployYieldRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PayrollJsonRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": "API payroll batch"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/PayrollRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "auto_execute": {
            "type": "boolean",
            "title": "Auto Execute",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "rows"
        ],
        "title": "PayrollJsonRequest"
      },
      "PayrollRow": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "amount_usd": {
            "type": "number",
            "title": "Amount Usd"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "default": ""
          },
          "country": {
            "type": "string",
            "title": "Country",
            "default": ""
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": ""
          },
          "chain": {
            "type": "string",
            "title": "Chain",
            "default": ""
          },
          "asset": {
            "type": "string",
            "title": "Asset",
            "default": "USDC"
          },
          "wallet_address": {
            "type": "string",
            "title": "Wallet Address",
            "default": ""
          },
          "prefers_fiat": {
            "type": "boolean",
            "title": "Prefers Fiat",
            "default": false
          },
          "memo": {
            "type": "string",
            "title": "Memo",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name",
          "amount_usd"
        ],
        "title": "PayrollRow"
      },
      "SwapRequest": {
        "properties": {
          "from_asset": {
            "type": "string",
            "title": "From Asset",
            "default": "USDC"
          },
          "to_asset": {
            "type": "string",
            "title": "To Asset",
            "default": "USDT"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "from_chain": {
            "type": "string",
            "title": "From Chain",
            "default": "base"
          },
          "to_chain": {
            "type": "string",
            "title": "To Chain",
            "default": "base"
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "SwapRequest"
      },
      "TransferRequest": {
        "properties": {
          "to_address": {
            "type": "string",
            "title": "To Address"
          },
          "amount_usd": {
            "type": "number",
            "title": "Amount Usd"
          },
          "asset": {
            "type": "string",
            "title": "Asset",
            "default": "USDC"
          },
          "chain": {
            "type": "string",
            "title": "Chain",
            "default": "base"
          },
          "counterparty": {
            "type": "string",
            "title": "Counterparty",
            "default": ""
          },
          "memo": {
            "type": "string",
            "title": "Memo",
            "default": ""
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "vendor"
          }
        },
        "type": "object",
        "required": [
          "to_address",
          "amount_usd"
        ],
        "title": "TransferRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}